+++ docs/changelog.md
@@ -21,6 +21,7 @@
 - Added sound file updates to include pop, q and others.
 - Added sounds for OIC and OICIC socials. These are in neuter until I can find ones to match to proper genders.
 - Added a couple of requested gags for bulk salvaging
+- Added sounds for bite, bap, howl, lick, thump and blink socials. 
 
 ### Fixed
 - Properly import timestamps from the database so older messages don't show up as "Just now".